+++ src/components/utils/search-input.ts
@@ -1,6 +1,5 @@
 import Dom from '../dom';
 import Listeners from './listeners';
-import $ from "../dom";
 
 /**
  * Item that could be searched
@@ -13,23 +12,47 @@ interface SearchableItem {
  * Provides search input element and search logic
  */
 export default class SearchInput {
-
+  /**
+   * Input wrapper element
+   */
   private wrapper: HTMLElement;
+
+  /**
+   * Editable input itself
+   */
   private input: HTMLInputElement;
+
+  /**
+   * The instance of the Listeners util
+   */
   private listeners: Listeners;
+
+  /**
+   * Items for local search
+   */
   private items: SearchableItem[];
+
+  /**
+   * Current search query
+   */
   private searchQuery: string;
-  private onSearch: (items: SearchableItem[]) => void;
+
+  /**
+   * Externally passed callback for the search
+   */
+  private readonly onSearch: (items: SearchableItem[]) => void;
 
   /**
    * Styles
    */
   private static get CSS(): {
     input: string;
+    icon: string;
     wrapper: string;
     } {
     return {
       wrapper: 'cdx-search-field',
+      icon: 'cdx-search-field__icon',
       input: 'cdx-search-field__input',
     };
   }
@@ -67,6 +90,15 @@ export default class SearchInput {
   }
 
   /**
+   * Clears search query and results
+   */
+  public clear(): void {
+    this.input.value = '';
+    this.searchQuery = '';
+    this.onSearch(this.foundItems);
+  }
+
+  /**
    * Clears memory
    */
   public destroy(): void {
@@ -80,13 +112,16 @@ export default class SearchInput {
    */
   private render(placeholder: string): void {
     this.wrapper = Dom.make('div', SearchInput.CSS.wrapper);
-    const icon = $.svg('search', 16, 16);
+
+    const iconWrapper = Dom.make('div', SearchInput.CSS.icon);
+    const icon = Dom.svg('search', 16, 16);
 
     this.input = Dom.make('input', SearchInput.CSS.input, {
       placeholder,
     }) as HTMLInputElement;
 
-    this.wrapper.appendChild(icon);
+    iconWrapper.appendChild(icon);
+    this.wrapper.appendChild(iconWrapper);
     this.wrapper.appendChild(this.input);
 
     this.listeners.on(this.input, 'input', () => {